Upload system

+++ app/api/admin/mainPosts/[id]/route.ts
@@ -2,6 +2,15 @@ import { getServerSession } from "next-auth/next"
 import { authOptions } from "@/lib/authOptions"
 import prisma from "@/lib/prisma"
 
+import { v2 as cloudinary } from 'cloudinary';
+
+cloudinary.config({
+    cloud_name: process.env.CLOUDNARY_CLOUD_NAME,
+    api_key: process.env.CLOUDNARY_API_KEY,
+    api_secret: process.env.CLOUDNARY_API_SECRET
+});
+
+
 export async function DELETE(
     request: Request,
     { params }: { params: { id: string } }
@@ -12,12 +21,20 @@ export async function DELETE(
     }
     const id = params.id
     try {
-        const res = await prisma.mainPost.delete({
+        const getPost = await prisma.mainPost.findUnique({
             where: {
                 id: id
             }
         })
-        return new Response(JSON.stringify(res), { status: 200 })
+        if (getPost) {
+            const deleteResponse = await cloudinary.uploader.destroy(getPost?.public_id)
+            const res = await prisma.mainPost.delete({
+                where: {
+                    id: id
+                }
+            })
+            return new Response(JSON.stringify(res), { status: 200 })
+        }
     } catch (error) {
         return new Response(JSON.stringify(error), { status: 500 })
     }
